Shows that current code is not properly validated
There is one test which is properly failing, the code of the application needs fixing.

Prevent duplication of code when validating the application against db.
