+++ src/test/java/io/jenkins/pluginhealth/scoring/PluginHealthScoringIT.java
@@ -24,45 +24,17 @@
 
 package io.jenkins.pluginhealth.scoring;
 
-import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.util.TestPropertyValues;
-import org.springframework.context.ApplicationContextInitializer;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-@ContextConfiguration(initializers = PluginHealthScoringIT.DockerPostgresDatasourceInitializer.class)
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@Testcontainers
-class PluginHealthScoringIT {
-    @Container
-    private static final PostgreSQLContainer POSTGRE_SQL_CONTAINER = new PostgreSQLContainer("postgres:14.1")
-        .withDatabaseName("testdb")
-        .withUsername("sa")
-        .withPassword("sa");
-
-    public static class DockerPostgresDatasourceInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
-        @Override
-        public void initialize(@NotNull ConfigurableApplicationContext applicationContext) {
-            TestPropertyValues.of(
-                "spring.datasource.url=" + POSTGRE_SQL_CONTAINER.getJdbcUrl(),
-                "spring.datasource.username=" + POSTGRE_SQL_CONTAINER.getUsername(),
-                "spring.datasource.password=" + POSTGRE_SQL_CONTAINER.getPassword()
-            ).applyTo(applicationContext);
-        }
-    }
-
+public final class PluginHealthScoringIT extends AbstractDBContainerTest {
     @Test
     void contextLoads() {
     }
-
 }
